deps: upgrade v8 to 4.1.0.7
This commit upgrades V8 from 3.31.74.1 to 4.1.0.7.  Despite the major
version bump, there are no API or ABI changes, it's a bug fix release
only.

PR-URL: https://github.com/iojs/io.js/pull/490

+++ deps/v8/src/d8.cc
@@ -1577,7 +1577,7 @@ class StartupDataHandler {
     const char* last_slash = strrchr(exec_path, '/');
     if (last_slash) {
       int after_slash = last_slash - exec_path + 1;
-      int name_length = strlen(name);
+      int name_length = static_cast<int>(strlen(name));
       *buffer =
           reinterpret_cast<char*>(calloc(after_slash + name_length + 1, 1));
       strncpy(*buffer, exec_path, after_slash);
